# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HabeshaShield API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{username}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get a user profile by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/profile:
    put:
      operationId: updateProfile
      tags: [users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/follow:
    post:
      operationId: followUser
      tags: [users]
      summary: Follow a user/artist
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Followed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      operationId: unfollowUser
      tags: [users]
      summary: Unfollow a user/artist
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unfollowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ARTISTS
...
  /tracks:
    get:
      operationId: listTracks
      tags: [tracks]
      summary: List tracks
      parameters:
        - name: genre
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: artistId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of tracks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackListResponse"
...
  /tracks/liked:
    get:
      operationId: getLikedTracks
      tags: [tracks]
      summary: Get tracks liked by current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Liked tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"
...
  /tracks/{id}:
    get:
      operationId: getTrack
      tags: [tracks]
      summary: Get a specific track
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Track details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
  /tracks/{id}/like:
    post:
      operationId: likeTrack
      tags: [tracks]
      summary: Like a track
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Liked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      operationId: unlikeTrack
      tags: [tracks]
      summary: Unlike a track
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unliked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /tracks/{id}/report:
    post:
      operationId: reportTrack
      tags: [tracks]
      summary: Report a track
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "200":
          description: Reported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # COMMENTS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # AUTH
...
    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        displayName:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, artist, admin]
        isVerified:
          type: boolean
        isBanned:
          type: boolean
        followersCount:
          type: integer
        followingCount:
          type: integer
        createdAt:
          type: string
    UserProfile:
      type: object
      required: [id, username, role, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        role:
          type: string
        isVerified:
          type: boolean
        followersCount:
          type: integer
        followingCount:
          type: integer
        tracksCount:
          type: integer
        isFollowing:
          type: boolean
        createdAt:
          type: string
    ProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: string
        avatar:
          type: string

    # ARTIST
...
    Track:
      type: object
      required: [id, title, artistId, artistName, genre, youtubeId, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        artistId:
          type: string
        artistName:
          type: string
        artistAvatar:
          type: ["string", "null"]
        genre:
          type: string
        youtubeId:
          type: string
        thumbnail:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        likesCount:
          type: integer
        commentsCount:
          type: integer
        isLiked:
          type: boolean
        isTrending:
          type: boolean
        createdAt:
          type: string
    TrackListResponse:
      type: object
      required: [tracks, total, page, totalPages]
      properties:
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
...
    ReportInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
        details:
          type: string
...